feat: extract style props and refine animation effect for the indeterminate state
